+++ CHANGELOG.rst
@@ -20,10 +20,12 @@ Features
 - C++:
 
   - ``storeChunk`` argument order changed, defaults added #386
+  - ``loadChunk`` argument order changed, defaults added #408
 - Python:
 
   - ``import openPMD`` renamed to ``import openpmd_api`` #380 #392
   - ``store_chunk`` argument order changed, defaults added #386
+  - ``load_chunk`` defaults added #408
   - works with Python 3.7 #376
   - setup.py for sdist #240
 - Backends: JSON support added #384 #393 #338